// src/api/generated/model/productImageResponseDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with Keycloak Authentication and Versioning
 * OpenAPI spec version: v1
 */
import type { ImageType } from './imageType';

export interface ProductImageResponseDto {
  id?: string;
  productId?: string;
  /** @nullable */
  cloudinaryPublicId?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  largeUrl?: string | null;
  imageType?: ImageType;
  displayOrder?: number;
  isActive?: boolean;
  fileSizeBytes?: number;
  /** @nullable */
  altText?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
}