// src/api/generated/model/imageType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with Keycloak Authentication and Versioning
 * OpenAPI spec version: v1
 */

export type ImageType = typeof ImageType[keyof typeof ImageType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ImageType = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
} as const;