// src/api/generated/model/postApiV1ProductIdImagesBody.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with Keycloak Authentication and Versioning
 * OpenAPI spec version: v1
 */
import type { ProductImageWebDto } from './productImageWebDto';

export type PostApiV1ProductIdImagesBody = {
  Images?: ProductImageWebDto[];
};