// src/api/generated/model/updateImageOrderRequest.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ECommerce API
 * ECommerce API with Keycloak Authentication and Versioning
 * OpenAPI spec version: v1
 */
import type { UpdateImageOrderRequestImageOrders } from './updateImageOrderRequestImageOrders';

export interface UpdateImageOrderRequest {
  /** @nullable */
  imageOrders?: UpdateImageOrderRequestImageOrders;
}